requirements.txt
matplotlib
numpy
opencv-python
pyyaml
flask
tensorboard
ultralytics